# openapi.yaml
openapi: 3.1.0
info:
  title: Reminders Service API
  version: 1.0.0
  description: |
    REST API for managing reminders in a Service-Oriented Architecture (SOA) ecosystem.
    Supports idempotent operations, RFC 7807 error responses, and OAuth2/OIDC security.
  contact:
    name: API Support
    email: support@reminders.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development

security:
  - bearerAuth: []
  - oauth2: [read:reminders, write:reminders]

tags:
  - name: Reminders
    description: Operations for managing reminders

paths:
  /reminders:
    post:
      tags:
        - Reminders
      summary: Create a new reminder
      description: |
        Creates a new reminder with idempotency support. 
        Repeated requests with the same Idempotency-Key will return the same result.
      operationId: createReminder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
            examples:
              basic:
                summary: Basic reminder
                value:
                  userId: "user-123"
                  title: "Submit assignment"
                  dueAt: "2025-11-15T10:00:00Z"
                  advanceMinutes: 30
              withSource:
                summary: Reminder from LMS
                value:
                  userId: "user-456"
                  title: "Exam tomorrow"
                  dueAt: "2025-11-20T09:00:00Z"
                  advanceMinutes: 60
                  source: "LMS"
                  metadata:
                    courseId: "CS101"
                    examType: "midterm"
      responses:
        '201':
          description: Reminder created successfully
          headers:
            Location:
              description: URI of the created reminder
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '200':
          description: Idempotent request - reminder already exists with this key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Reminders
      summary: List reminders
      description: Retrieve a paginated list of reminders for the authenticated user
      operationId: listReminders
      parameters:
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by reminder status
          schema:
            $ref: '#/components/schemas/ReminderStatus'
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of reminders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /reminders/{id}:
    parameters:
      - $ref: '#/components/parameters/ReminderId'
    
    get:
      tags:
        - Reminders
      summary: Get a reminder by ID
      description: Retrieve a specific reminder by its unique identifier
      operationId: getReminderById
      responses:
        '200':
          description: Reminder found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    patch:
      tags:
        - Reminders
      summary: Update a reminder
      description: Partially update an existing reminder
      operationId: updateReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
            examples:
              updateTitle:
                summary: Update title only
                value:
                  title: "Updated reminder title"
              updateStatus:
                summary: Mark as completed
                value:
                  status: "completed"
              reschedule:
                summary: Reschedule reminder
                value:
                  dueAt: "2025-11-16T10:00:00Z"
                  advanceMinutes: 45
      responses:
        '200':
          description: Reminder updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Reminders
      summary: Delete a reminder
      description: Soft delete a reminder (marks as cancelled)
      operationId: deleteReminder
      responses:
        '204':
          description: Reminder deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/oauth2/authorize
          tokenUrl: https://auth.example.com/oauth2/token
          scopes:
            read:reminders: Read access to reminders
            write:reminders: Write access to reminders

  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Unique identifier (UUID v4) for idempotent operations. 
        Repeated requests with the same key will return the same result.
      required: true
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      example: "550e8400-e29b-41d4-a716-446655440000"
    
    ReminderId:
      name: id
      in: path
      description: Unique identifier of the reminder
      required: true
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"

  schemas:
    Reminder:
      type: object
      required:
        - id
        - userId
        - title
        - dueAt
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the reminder
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          description: ID of the user who owns this reminder
          example: "user-123"
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Title or description of the reminder
          example: "Submit assignment"
        dueAt:
          type: string
          format: date-time
          description: When the reminder is due (ISO 8601)
          example: "2025-11-15T10:00:00Z"
        status:
          $ref: '#/components/schemas/ReminderStatus'
        source:
          type: string
          enum: [manual, LMS, calendar, external]
          description: Source system that created the reminder
          default: manual
          example: "manual"
        advanceMinutes:
          type: integer
          minimum: 0
          maximum: 10080
          description: How many minutes before dueAt to send notification (max 7 days)
          default: 15
          example: 30
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata from external systems
          example:
            courseId: "CS101"
            assignmentType: "homework"
        notifiedAt:
          type: string
          format: date-time
          nullable: true
          description: When the notification was sent
          example: "2025-11-15T09:30:00Z"
        createdAt:
          type: string
          format: date-time
          description: When the reminder was created
          example: "2025-11-10T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the reminder was last updated
          example: "2025-11-10T08:00:00Z"

    ReminderStatus:
      type: string
      enum:
        - pending
        - scheduled
        - notified
        - completed
        - cancelled
      description: |
        Current status of the reminder:
        - pending: Created but not yet scheduled
        - scheduled: Scheduled in orchestration engine
        - notified: Notification sent
        - completed: User marked as done
        - cancelled: User cancelled
      example: "pending"

    CreateReminderRequest:
      type: object
      required:
        - userId
        - title
        - dueAt
      properties:
        userId:
          type: string
          description: ID of the user who owns this reminder
          example: "user-123"
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Title or description of the reminder
          example: "Submit assignment"
        dueAt:
          type: string
          format: date-time
          description: When the reminder is due (ISO 8601)
          example: "2025-11-15T10:00:00Z"
        source:
          type: string
          enum: [manual, LMS, calendar, external]
          description: Source system that created the reminder
          default: manual
        advanceMinutes:
          type: integer
          minimum: 0
          maximum: 10080
          description: Minutes before dueAt to notify (max 7 days)
          default: 15
          example: 30
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata
          example:
            courseId: "CS101"

    UpdateReminderRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated title
        dueAt:
          type: string
          format: date-time
          description: Updated due date
        status:
          $ref: '#/components/schemas/ReminderStatus'
        advanceMinutes:
          type: integer
          minimum: 0
          maximum: 10080
          description: Updated advance notification time
        metadata:
          type: object
          additionalProperties: true
          description: Updated metadata

    ReminderList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Reminder'
        pagination:
          type: object
          required:
            - page
            - limit
            - total
            - totalPages
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 42
            totalPages:
              type: integer
              example: 3

    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: "https://api.example.com/problems/validation-error"
        title:
          type: string
          description: A short, human-readable summary of the problem
          example: "Validation Error"
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: "The field 'dueAt' must be a future date"
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: "/v1/reminders"
        errors:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: "dueAt"
              message:
                type: string
                example: "must be a future date"
              code:
                type: string
                example: "INVALID_DATE"
        traceId:
          type: string
          description: Trace ID for distributed tracing
          example: "a1b2c3d4e5f6"

  responses:
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            validationError:
              summary: Validation error
              value:
                type: "https://api.example.com/problems/validation-error"
                title: "Validation Error"
                status: 400
                detail: "Request validation failed"
                instance: "/v1/reminders"
                errors:
                  - field: "dueAt"
                    message: "must be a future date"
                    code: "INVALID_DATE"
                  - field: "advanceMinutes"
                    message: "must be between 0 and 10080"
                    code: "OUT_OF_RANGE"
                traceId: "a1b2c3d4e5f6"
            invalidFormat:
              summary: Invalid format
              value:
                type: "https://api.example.com/problems/invalid-format"
                title: "Invalid Format"
                status: 400
                detail: "The request body is not valid JSON"
                instance: "/v1/reminders"
                traceId: "b2c3d4e5f6a7"

    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.example.com/problems/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Authentication token is missing or invalid"
            instance: "/v1/reminders"
            traceId: "c3d4e5f6a7b8"

    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.example.com/problems/forbidden"
            title: "Forbidden"
            status: 403
            detail: "You don't have permission to access this resource"
            instance: "/v1/reminders/123e4567-e89b-12d3-a456-426614174000"
            traceId: "d4e5f6a7b8c9"

    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.example.com/problems/not-found"
            title: "Not Found"
            status: 404
            detail: "Reminder with ID '123e4567-e89b-12d3-a456-426614174000' not found"
            instance: "/v1/reminders/123e4567-e89b-12d3-a456-426614174000"
            traceId: "e5f6a7b8c9d0"

    Conflict:
      description: Conflict - Idempotency key conflict or resource conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            idempotencyConflict:
              summary: Idempotency key conflict
              value:
                type: "https://api.example.com/problems/idempotency-conflict"
                title: "Idempotency Conflict"
                status: 409
                detail: "A different request with the same Idempotency-Key was previously processed"
                instance: "/v1/reminders"
                traceId: "f6a7b8c9d0e1"
            resourceConflict:
              summary: Resource state conflict
              value:
                type: "https://api.example.com/problems/conflict"
                title: "Conflict"
                status: 409
                detail: "Cannot update reminder in 'notified' status"
                instance: "/v1/reminders/123e4567-e89b-12d3-a456-426614174000"
                traceId: "a7b8c9d0e1f2"

    InternalServerError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.example.com/problems/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred while processing your request"
            instance: "/v1/reminders"
            traceId: "b8c9d0e1f2a3"